begun implementing manual BPM input
this doesn't work yet, but the idea is you can input the BPM you want
manually. It's bounded by 60 and 240 bpm. If you hit enter it sets BPM
to whatever is in the buffer, and if you type 3 numbers in it
automaticalyl sets the BPM to whatever is in the buffer and clears it.